address e2e review comment
