+++ workspaces/tekton/plugins/tekton-dev-backend/src/index.ts
@@ -14,7 +14,20 @@
  * limitations under the License.
  */
 
-/**
- * @internal
- */
-export { tektonDevBackendPlugin as default } from './plugin';
+import { createBackend } from '@backstage/backend-defaults';
+
+const backend = createBackend();
+
+backend.add(import('@backstage/plugin-auth-backend'));
+backend.add(import('@backstage/plugin-auth-backend-module-guest-provider'));
+
+backend.add(import('@backstage/plugin-catalog-backend'));
+
+backend.add(import('@backstage/plugin-permission-backend'));
+backend.add(
+  import('@backstage/plugin-permission-backend-module-allow-all-policy'),
+);
+
+backend.add(import('@backstage/plugin-kubernetes-backend'));
+
+backend.start();
