--- workspaces/tekton/plugins/tekton-dev-backend/src/plugin.ts
@@ -1,34 +1,0 @@
-/*
- * Copyright 2026 The Backstage Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { createBackendPlugin } from '@backstage/backend-plugin-api';
-
-/**
- * Minimal backend plugin pairs this package with
- * `@backstage-community/plugin-tekton`.
- * Dev-only wiring (catalog, auth, kubernetes, …) lives in `dev/index.ts`.
- *
- * @internal
- */
-export const tektonDevBackendPlugin = createBackendPlugin({
-  pluginId: 'tekton',
-  register(env) {
-    env.registerInit({
-      deps: {},
-      async init() {},
-    });
-  },
-});
